requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
meshio==5.2.5
numpy==1.22.1
Pygments==2.11.2
pyzmq==22.3.0
requests==2.27.1
rich==11.0.0
urllib3==1.26.8
Werkzeug==2.0.2
